requirements.txt
cryptography==3.3.2
mitmproxy==6.0.2
Pillow==8.2.0
python-dotenv==0.17.0
requests==2.25.1
selenium==3.141.0
selenium-wire==4.2.3
undetected-chromedriver==2.2.1
